Add Postgres get methods and update demo
This commits disable the Drill processor that crashes the application, to re-enable the processor uncomment the line `this.drillProcessor = new DrillQueryProcessor(this);` in DemoMiddleware.
Note that this crash may be due to the addition of the Postgres JDBC driver in the classpath, or because of network issues, and need to be investigated.